backend/requirements.txt
fastapi[standard]==0.116.1
uvicorn==0.33.0
python-multipart==0.0.18
pydantic-settings==2.7.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
retell-sdk==4.42.0
supabase==2.9.1
